+++ EnchantmentsEnhance-Core/src/main/kotlin/org/pixeltime/enchantmentsenhance/event/enchantment/gear/Paralyze.kt
@@ -28,7 +28,6 @@ import org.bukkit.event.entity.EntityDamageByEntityEvent
 import org.bukkit.potion.PotionEffect
 import org.bukkit.potion.PotionEffectType
 import org.pixeltime.enchantmentsenhance.listener.EnchantmentListener
-import org.pixeltime.enchantmentsenhance.manager.IM
 import org.pixeltime.enchantmentsenhance.manager.SettingsManager
 
 class Paralyze : EnchantmentListener() {
@@ -54,7 +53,7 @@ class Paralyze : EnchantmentListener() {
             }
 
             try {
-                val level =getLevel(player)
+                val level = getLevel(player)
                 if (level > 0 && (roll(level))) {
                     player2.addPotionEffect(PotionEffect(PotionEffectType.getById(4), SettingsManager.enchant.getInt("paralyze.$level.duration") * 20, SettingsManager.enchant.getInt("paralyze.$level.potion_lvl") - 1))
                 }